Move wide decimal compat coverage into its own fixture
The wide `DecimalByteParts` columns were added to the existing
`decimal_byte_parts.vortex` fixture, which breaks the compat contract.
`DESIGN.md` states it directly under "Fixture evolution": a fixture's
`build()` is immutable once published, because `check` compares files
written by older releases against what `build()` produces today. Adding a
column changes the schema the generator emits, so the check fails against
every previously published version — exactly the regression the fixture
exists to catch, reported against unrelated releases.

`decimal_byte_parts.vortex` is restored to its published definition, and
the wide cases move to a new `decimal_byte_parts_wide.vortex` with a
comment recording why the split exists rather than leaving the next person
to rediscover the rule. The new fixture gains a negative `i128` column so
sign extension above the MSP is exercised on read back, alongside the
one-lower-part and nullable three-lower-part cases.

Verified with `generate` followed by `check --mode exact`: 36 fixtures
pass, and `decimal_byte_parts.rs` is byte-identical to its pre-branch
state.
